feat: Add retry with backoff to requests
This adds a bit more tolerance to network outages etc. If a 40x status code is recieved from the API
it will not retry as this indicates a bad request – so no need to keep sening the same
request.

There were no end-to-end tests predating this feature so I added some.

Fixes https://github.com/bugsnag/webpack-bugsnag-plugins/issues/17.