feat: Retry requests in more situations
This flips the retry logic from a flag that says "retry" me, to a flag that says "don't retry me" –
and only if the "don't retry me" flag exists do we not retry. A couple of test cases are added for
new cases where we now retry, such as socket timeouts and socket hangups.